# docs/api-documentation.yaml
openapi: 3.0.3
info:
  title: ActionLoom AI Agent API
  description: |
    ActionLoom AI Agent API provides programmatic access to blockchain workflow composition, 
    Action discovery, and Agent management capabilities. This API is designed for AI agents, 
    trading bots, and automated systems to integrate with Flow blockchain automation.
    
    ## Authentication
    All API endpoints require authentication using API keys. Include your API key in the 
    Authorization header as a Bearer token:
    
    ```
    Authorization: Bearer al_your_api_key_here
    ```
    
    ## Rate Limits
    API requests are rate-limited based on your subscription tier:
    
    - **Free Tier**: 10 requests/minute, 100 requests/hour, 1,000 requests/day
    - **Pro Tier**: 100 requests/minute, 5,000 requests/hour, 50,000 requests/day  
    - **Enterprise Tier**: 1,000 requests/minute, 50,000 requests/hour, 1,000,000 requests/day
    
    Rate limit information is included in response headers:
    - `X-RateLimit-Remaining-Minute`: Remaining requests this minute
    - `X-RateLimit-Remaining-Hour`: Remaining requests this hour
    - `X-RateLimit-Remaining-Day`: Remaining requests today
    
    ## Error Handling
    The API uses standard HTTP status codes and returns consistent error responses:
    
    ```json
    {
      "success": false,
      "error": "Error description",
      "details": "Additional error details"
    }
    ```
    
    Common status codes:
    - `200`: Success
...
    
...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /agents:
    get:
      summary: List Agents
      description: |
        List blockchain automation Agents. Can filter by user, status, and include 
        health information.
      operationId: listAgents
      tags:
        - Agent Management
      parameters:
        - name: id
          in: query
          description: Get specific Agent by ID
          schema:
            type: string
        - name: userId
          in: query
          description: Filter Agents by user ID
          schema:
            type: string
        - name: status
          in: query
          description: Filter Agents by status
          schema:
            type: string
            enum: [active, paused, stopped, error]
        - name: health
          in: query
          description: Include Agent health information
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Agents listed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentsResponse'
        '404':
          description: Agent not found (when requesting specific Agent by ID)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create Agent
      description: |
        Create a new blockchain automation Agent from a workflow. The Agent will 
        execute the workflow according to the specified schedule and triggers.
      operationId: createAgent
      tags:
        - Agent Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
            examples:
              fromWorkflow:
                summary: Create Agent from Workflow
                value:
                  workflow:
                    actions:
                      - id: "swap1"
                        actionType: "swap"
                        name: "Daily DCA"
                        parameters:
                          - name: "fromToken"
                            type: "string"
                            value: "USDC"
                            required: true
                          - name: "toToken"
                            type: "string"
                            value: "FLOW"
                            required: true
                          - name: "amount"
                            type: "number"
                            value: "10"
                            required: true
                        nextActions: []
                        position:
                          x: 0
                          y: 0
                    executionOrder: ["swap1"]
                    rootActions: ["swap1"]
                    metadata:
                      totalActions: 1
                      totalConnections: 0
                      createdAt: "2024-01-01T00:00:00Z"
                  config:
                    schedule:
                      type: "recurring"
                      interval: 86400
                    eventTriggers: []
                    retryPolicy:
                      maxRetries: 3
                      backoffStrategy: "exponential"
                      initialDelay: 1000
                      maxDelay: 10000
                    notifications: {}
                    permissions: []
                  name: "Daily DCA Agent"
                  description: "Automated daily dollar-cost averaging"
                  owner: "user123"
      responses:
        '200':
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAgentResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update Agent
      description: Update an existing Agent's configuration, schedule, or metadata.
      operationId: updateAgent
      tags:
        - Agent Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
      responses:
        '200':
          description: Agent updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateAgentResponse'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete Agent
      description: Delete an Agent and stop all its scheduled executions.
      operationId: deleteAgent
      tags:
        - Agent Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAgentRequest'
      responses:
        '200':
          description: Agent deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAgentResponse'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /agents/control:
    post:
      summary: Control Agent
      description: Control Agent execution (pause, resume, stop).
      operationId: controlAgent
      tags:
        - Agent Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentControlRequest'
      responses:
        '200':
          description: Agent control action completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentControlResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        API key authentication. Include your API key in the Authorization header:
        `Authorization: Bearer al_your_api_key_here`
  schemas:
...
    ActionParameter:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        value:
          type: string
        required:
          type: boolean
...
    AgentsResponse:
      type: object
      properties:
        success:
          type: boolean
        agents:
          type: array
          items:
            type: object
            properties:
              agent:
                $ref: '#/components/schemas/Agent'
              health:
                $ref: '#/components/schemas/AgentHealth'
        total:
          type: integer
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        workflowId:
          type: string
        schedule:
          $ref: '#/components/schemas/Schedule'
        triggers:
          type: array
          items:
            $ref: '#/components/schemas/EventTrigger'
        status:
          type: string
          enum: [active, paused, stopped, error]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        owner:
          type: string
    AgentHealth:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
        uptime:
          type: integer
          description: Uptime in seconds
        lastExecution:
          type: string
          format: date-time
        successRate:
          type: number
          format: float
          minimum: 0
          maximum: 1
        errorCount:
          type: integer
        lastError:
          type: string
    CreateAgentRequest:
      type: object
      required: [config, name, owner]
      properties:
        workflow:
          $ref: '#/components/schemas/ParsedWorkflow'
        workflowId:
          type: string
        config:
          $ref: '#/components/schemas/AgentConfiguration'
        name:
          type: string
        description:
          type: string
        owner:
          type: string
      oneOf:
        - required: [workflow]
        - required: [workflowId]
    CreateAgentResponse:
      type: object
      properties:
        success:
          type: boolean
        agentId:
          type: string
        agent:
          $ref: '#/components/schemas/Agent'
        message:
          type: string
    UpdateAgentRequest:
      type: object
      required: [agentId, updates]
      properties:
        agentId:
          type: string
        updates:
          type: object
          properties:
            schedule:
              $ref: '#/components/schemas/Schedule'
            eventTriggers:
              type: array
              items:
                $ref: '#/components/schemas/EventTrigger'
            retryPolicy:
              $ref: '#/components/schemas/RetryPolicy'
            notifications:
              $ref: '#/components/schemas/NotificationConfig'
            permissions:
              type: array
              items:
                $ref: '#/components/schemas/Permission'
            name:
              type: string
            description:
              type: string
    UpdateAgentResponse:
      type: object
      properties:
        success:
          type: boolean
        agent:
          $ref: '#/components/schemas/Agent'
        message:
          type: string
    DeleteAgentRequest:
      type: object
      required: [agentId]
      properties:
        agentId:
          type: string
    DeleteAgentResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    AgentControlRequest:
      type: object
      required: [agentId, action]
      properties:
        agentId:
          type: string
        action:
          type: string
          enum: [pause, resume, stop]
    AgentControlResponse:
      type: object
      properties:
        success:
          type: boolean
        agent:
          $ref: '#/components/schemas/Agent'
        message:
          type: string
...
    AgentConfiguration:
      type: object
      required: [schedule, eventTriggers, retryPolicy, notifications, permissions]
      properties:
        schedule:
          $ref: '#/components/schemas/Schedule'
        eventTriggers:
          type: array
          items:
            $ref: '#/components/schemas/EventTrigger'
        retryPolicy:
          $ref: '#/components/schemas/RetryPolicy'
        notifications:
          $ref: '#/components/schemas/NotificationConfig'
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
    Schedule:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [recurring, event-driven, one-time]
        interval:
          type: integer
          description: Interval in seconds for recurring schedules
        cronExpression:
          type: string
          description: Cron expression for complex schedules
        eventTriggers:
          type: array
          items:
            $ref: '#/components/schemas/EventTrigger'
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
    EventTrigger:
      type: object
      required: [type, condition, parameters]
      properties:
        type:
          type: string
          enum: [price, time, balance, custom]
        condition:
          $ref: '#/components/schemas/TriggerCondition'
        oracleAction:
          type: string
          description: Action ID for oracle data
        parameters:
          type: object
          additionalProperties: true
    TriggerCondition:
      type: object
      required: [operator, value]
      properties:
        operator:
          type: string
          enum: [gt, lt, eq, gte, lte]
        value:
          oneOf:
            - type: string
            - type: number
            - type: boolean
        tolerance:
          type: number
    RetryPolicy:
      type: object
      required: [maxRetries, backoffStrategy, initialDelay, maxDelay]
      properties:
        maxRetries:
          type: integer
          minimum: 0
          maximum: 10
        backoffStrategy:
          type: string
          enum: [linear, exponential]
        initialDelay:
          type: integer
          description: Initial delay in milliseconds
        maxDelay:
          type: integer
          description: Maximum delay in milliseconds
    NotificationConfig:
      type: object
      properties:
        email:
          type: string
          format: email
        webhook:
          type: string
          format: uri
        discord:
          type: string
        slack:
          type: string
    Permission:
      type: object
      required: [resource, action]
      properties:
        resource:
          type: string
        action:
          type: string
        conditions:
          type: object
          additionalProperties: true

    # Workflow schemas
    ParsedWorkflow:
      type: object
      required: [actions, executionOrder, rootActions, metadata]
      properties:
        actions:
          type: array
          items:
            $ref: '#/components/schemas/ParsedAction'
        executionOrder:
          type: array
          items:
            type: string
        rootActions:
          type: array
          items:
            type: string
        metadata:
          $ref: '#/components/schemas/WorkflowMetadata'
...
    ParsedAction:
      type: object
      required: [id, actionType, name, parameters, nextActions, position]
      properties:
        id:
          type: string
        actionType:
          type: string
        name:
          type: string
        parameters:
          type: array
          items:
            $ref: '#/components/schemas/ActionParameter'
        nextActions:
          type: array
          items:
            type: string
        position:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
    WorkflowMetadata:
      type: object
      required: [totalActions, totalConnections, createdAt]
      properties:
        totalActions:
          type: integer
        totalConnections:
          type: integer
        createdAt:
          type: string
          format: date-time
        name:
          type: string
        savedAt:
          type: string
          format: date-time
        version:
          type: string
        author:
          type: string
        tags:
          type: array
          items:
            type: string

    # Simulation and execution schemas
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
          description: Human-readable error message
        details:
          type: string
          description: Additional technical details about the error
        code:
          type: string
          description: Machine-readable error code
...